+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## Next Version
 * Add support for supports_fraud_review gateway characteristic [markabe]
 * Add support for the store API call [duff]
+* Add support for company name property on credit card [duff]
 
 ## 2.0.14
 * Add support for gateway_specific_fields [duff]