+++ HISTORY.md
@@ -5,6 +5,7 @@
 * Add support for the store API call [duff]
 * Add support for company name property on credit card [duff]
 * Deliver payment method uses talking to gateway timeout [duff]
+* Fix issue creating production receiver [duff]
 
 ## 2.0.14
 * Add support for gateway_specific_fields [duff]